Styles, Attributes and Classes
**Style->
select_element>style>default_attribute

For undefined style we use
getComputedStyle- sometimes we did not define a style attribute , but automatically it applied on browser , but later on if we need to change it , we can use it

setProperty- Custom property for CSS

**Attribute->
We can not directly select an attribute.In this case we will need to select the element first. Remember we can only use default attributes.

getAttribute->To use a custom attributes we need to use getAttribute
setAttribute(_ , _)->

data Attribute- If we wrote in html data-version-number then in script we select element then dataset.versionNumber ( name needs to be camel case), remember the naming convention between html and js

**Classes-
element.classList.add
element.classList.remove
element.classList.toggle
element.classList.contains