Object.create
So when we use the new operator in constructor functions or classes, it automatically sets the prototype of the instances to the constructors, prototype property.

So this happens automatically. And so that's nothing new at this point .

Now, on the other hand, with Object.create, we can set the prototype of objects manually to any object that we want.

And in this case, we manually set the prototype of the Shafinul object to the person proto object.

The big difference is that we didn't need any constructor function, and also no prototype property at all, to achieve the exact same thing.

This is actually the least used way of implementing prototypal inheritance.