Inheritance Between Classes: Constructor Functions
1. Instead of having duplicate code here, let's simply call the person function.But this doesn't work.
So the problem here is that we are now actually calling this person constructor function as a regular function call. So we are not using this new operator to call this person function constructor. And so therefore this function call here is simply a regular function call.
And remember that in a regular function call, the this keyword is set to undefined.

So instead of simply calling the person function here, we need to manually set the this keyword as well.
Well, we can simply use the call method. So the call method will indeed call this function, but we will be able to specify the this keywords here as the first argument in this function.
Because as we know the this Keyword is gonna be in the beginning, this empty object that is being created by the new operator. And so it is on that new object where we want to set the first name and the birth year property.

2. Linking prototypes

3. Checking all the inherit prototypes