Encapsulation: Protected Properties and Methods
Encapsulation basically means to keep some properties and methods private inside the class so that they are not accessible from outside of the class.

Then the rest of the methods are basically exposed as a public interface, which we can also call API.

There are two big reasons why we need encapsulation and data privacy. So first it is to prevent code from outside of a class to accidentally manipulate or data inside the class. And that's exactly what we did by the end of the last lecture here. So this is also the reason why we implement a public interface. So we are not supposed to manually mess with this property and therefore we should encapsulate it.

Now, the second reason is that when we expose only a small interface so a small API consisting only of a few public methods then we can change all the other internal methods with more confidence. Because in this case, we can be sure that external code does not rely on these private methods. And so therefore our code will not break when we do internal changes. So that's what encapsulation and data privacy are and the reasons for it.

1. putting underscore. again this does not actually make the property truly private because this is just a convention.So it's something that developers agree to use

Codebase comments:

1.Protected property(putting underscore). again this does not actually make the property truly private because this is just a convention.So it's something that developers agree to use.

2.  Adding other methods to the class ------Public API's

3. Everyone can still at least access the movements but they cannot override them.